+++ caffe2/CMakeLists.txt
@@ -53,6 +53,7 @@ if(BUILD_CAFFE2)
   add_subdirectory(utils)
   add_subdirectory(contrib)
   add_subdirectory(core/nomnigraph)
+  add_subdirectory(core/dispatch)
   add_subdirectory(cuda_rtc)
   add_subdirectory(db)
   add_subdirectory(distributed)
@@ -175,7 +176,6 @@ if(BUILD_CAFFE2)
 endif()
 
 # Compile exposed libraries.
-list(APPEND Caffe2_CPU_SRCs $<TARGET_OBJECTS:c10>)
 add_library(caffe2 ${Caffe2_CPU_SRCS})
 if (BUILD_CAFFE2)
   caffe2_interface_library(caffe2_protos caffe2_protos_whole)
