+++ caffe2/core/dispatch/dummy.cpp
@@ -1,3 +1,3 @@
 // This is going to be replaced with actual c10 files
 
-#include <c10/dummy.h>
+#include "dummy.h"
